+++ src/shared_utilities/drivers.jl
@@ -25,11 +25,9 @@ export AbstractAtmosphericDrivers,
     CoupledAtmosphere,
     PrescribedRadiativeFluxes,
     CoupledRadiativeFluxes,
-    set_atmos_ts!,
     turbulent_fluxes!,
     net_radiation!,
     turbulent_fluxes_at_a_point,
-    vapor_pressure_deficit,
     make_update_drivers,
     prescribed_forcing_era5,
     prescribed_perturbed_temperature_era5,
@@ -318,39 +316,27 @@ struct CoupledAtmosphere{FT} <: AbstractAtmosphericDrivers{FT}
 end
 
 """
-    compute_ρ_sfc(thermo_params, ts_in, T_sfc)
+    compute_ρ_sfc(surface_flux_params, T_air, P_air, q_air, Δh, T_sfc)
 
 Computes the density of air at the surface, given the temperature
-at the surface T_sfc, the thermodynamic state of the atmosphere,
-ts_in, and a set of Clima.Thermodynamics parameters thermo_params.
-
-This assumes the ideal gas law and hydrostatic balance to
-extrapolate to the surface.
-
+at the surface T_sfc, the atmospheric temperature T_air, pressure P_air,
+specific humidity q_air, , and height difference between the atmos level
+and surface level, using surface_flux_params.
 """
-function compute_ρ_sfc(thermo_params, ts_in, T_sfc)
-    T_int = Thermodynamics.air_temperature(thermo_params, ts_in)
-    Rm_int = Thermodynamics.gas_constant_air(thermo_params, ts_in)
-    ρ_air = Thermodynamics.air_density(thermo_params, ts_in)
+function compute_ρ_sfc(surface_flux_params, T_air, P_air, q_air, Δh, T_sfc)
     T_sfc = T_sfc < 0 ? eltype(T_sfc)(NaN) : T_sfc
-    ρ_sfc =
-        ρ_air *
-        (T_sfc / T_int)^(Thermodynamics.cv_m(thermo_params, ts_in) / Rm_int)
+    thermo_params =
+        SurfaceFluxes.Parameters.thermodynamics_params(surface_flux_params)
+    ρ_air = Thermodynamics.air_density(thermo_params, T_air, P_air, q_air)
+    return SurfaceFluxes.surface_density(
+        surface_flux_params,
+        T_air,
+        ρ_air,
+        T_sfc,
+        Δh,
+        q_air,
+    )
     return ρ_sfc
-end
-
-"""
-    set_atmos_ts!(ts_in, atmos::PrescribedAtmosphere{FT}, p)
-
-Fill the pre-allocated ts_in `Field` with a thermodynamic state computed from the
-atmosphere.
-"""
-function set_atmos_ts!(ts_in, atmos::PrescribedAtmosphere{FT}, p) where {FT}
-    P = p.drivers.P
-    T = p.drivers.T
-    q = p.drivers.q
-    ts_in .= Thermodynamics.PhaseEquil_pTq.(atmos.thermo_params, P, T, q)
-    return nothing
 end
 
 return_momentum_fluxes(atmos::PrescribedAtmosphere) = false
@@ -511,19 +497,8 @@ function compute_turbulent_fluxes_at_a_point(
     else
         u = u_atmos
     end
-    phase_partition_atmos = Thermodynamics.PhasePartition_equil_given_p(
-        thermo_params,
-        T_atmos,
-        P_atmos,
-        q_tot_atmos,
-        Thermodynamics.PhaseEquil,
-    )
-    ρ_atmos = Thermodynamics.air_density(
-        thermo_params,
-        T_atmos,
-        P_atmos,
-        phase_partition_atmos,
-    )
+    ρ_atmos =
+        Thermodynamics.air_density(thermo_params, T_atmos, P_atmos, q_tot_atmos)
     output = SurfaceFluxes.surface_fluxes(
         surface_flux_params,
         T_atmos,
@@ -550,20 +525,11 @@ function compute_turbulent_fluxes_at_a_point(
     E = output.evaporation
 
     # vapor flux in volume of liquid water
-    Ẽ = E / _ρ_liq
+    Ẽ = E / _ρ_liq
 
-    # Approximate derivatives of fluxes with respect to T_sfc, q_sfc
+    # Approximate derivatives of fluxes with respect to T_sfc
     g_h = output.g_h
     u_star = output.ustar
-    ρ_sfc = ρ_atmos
-    ∂lhf∂T =
-        ρ_sfc *
-        g_h *
-        _LH_v0 *
-        update_∂q_sfc∂T(u_star, g_h, T_sfc_guess, P_atmos, earth_param_set)
-    cp_d = Thermodynamics.Parameters.cp_d(thermo_params)
-    ∂shf∂T = ρ_sfc * g_h * cp_d * update_∂T_sfc∂T(u_star, g_h, earth_param_set)
-    # Buoyancy Flux
     ρ_sfc = SurfaceFluxes.surface_density(
         surface_flux_params,
         T_atmos,
@@ -575,6 +541,21 @@ function compute_turbulent_fluxes_at_a_point(
         FT(0),
         output.q_vap_sfc,
     )
+    # This assumes that q_vap_sfc_guess is the saturated value
+    ∂lhf∂T =
+        ρ_sfc *
+        g_h *
+        _LH_v0 *
+        update_∂q_sfc∂T(
+            u_star,
+            g_h,
+            q_vap_sfc_guess,
+            T_sfc_guess,
+            earth_param_set,
+        )
+    cp_d = Thermodynamics.Parameters.cp_d(thermo_params)
+    ∂shf∂T = ρ_sfc * g_h * cp_d * update_∂T_sfc∂T(u_star, g_h, earth_param_set)
+    # Buoyancy Flux
     buoyancy_flux = SurfaceFluxes.buoyancy_flux(
         surface_flux_params,
         output.shf,
@@ -588,7 +569,7 @@ function compute_turbulent_fluxes_at_a_point(
     return (
         output.lhf,
         output.shf,
-        Ẽ,
+        Ẽ,
         ∂lhf∂T,
         ∂shf∂T,
         output.ρτxz,
@@ -833,13 +814,22 @@ function get_∂q_sfc∂T_function(model::AbstractModel, Y, p)
     function update_∂q_sfc∂T_at_a_point(
         u_star,
         g_h,
+        q_sat,
         T_sfc,
-        P_sfc,
         earth_param_set,
     )
-        FT = eltype(earth_param_set)
         _T_freeze = LP.T_freeze(earth_param_set)
-        return ClimaLand.partial_q_sat_partial_T(P_sfc, T_sfc - _T_freeze)
+        phase = ifelse(
+            T_sfc > _T_freeze,
+            Thermodynamics.Liquid(),
+            Thermodynamics.Ice(),
+        )
+        return ClimaLand.partial_q_sat_partial_T(
+            q_sat,
+            T_sfc,
+            phase,
+            earth_param_set,
+        )
     end
     return update_∂q_sfc∂T_at_a_point
 end
@@ -887,53 +877,6 @@ the functions in this file.
 function surface_emissivity(model::AbstractModel, Y, p) end
 
 """
-    vapor_pressure_deficit(T_air, P_air, q_air, thermo_params)
-
-Computes the vapor pressure deficit for air with temperature T_air,
-pressure P_air, and specific humidity q_air, using thermo_params,
-a Thermodynamics.jl param set.
-"""
-function vapor_pressure_deficit(T_air, P_air, q_air, thermo_params)
-    es = Thermodynamics.saturation_vapor_pressure(
-        thermo_params,
-        T_air,
-        Thermodynamics.Liquid(),
-    )
-    ea = Thermodynamics.partial_pressure_vapor(
-        thermo_params,
-        P_air,
-        Thermodynamics.PhasePartition(q_air),
-    )
-    return es - ea
-end
-
-"""
-    relative_humidity(T_air, P_air, q_air, thermo_params)
-
-Computes the vapor pressure deficit for air with temperature T_air,
-pressure P_air, and specific humidity q_air, using thermo_params,
-a Thermodynamics.jl param set.
-"""
-function relative_humidity(
-    T_air::FT,
-    P_air::FT,
-    q_air::FT,
-    thermo_params,
-) where {FT}
-    es = Thermodynamics.saturation_vapor_pressure(
-        thermo_params,
-        T_air,
-        Thermodynamics.Liquid(),
-    )
-    ea = Thermodynamics.partial_pressure_vapor(
-        thermo_params,
-        P_air,
-        Thermodynamics.PhasePartition(q_air),
-    )
-    return es / ea
-end
-
-"""
     specific_humidity_from_dewpoint(dewpoint_temperature, temperature, air_pressure, earth_param_set)
 
 Estimates the specific humidity given the dewpoint temperature, temperature of the air
@@ -963,11 +906,12 @@ function specific_humidity_from_dewpoint(
         sim_FT(T_dew_air) - _T_freeze,
         sim_FT(T_air) - _T_freeze,
     )
-    q = Thermodynamics.q_vap_from_RH_liquid(
+    q = Thermodynamics.q_vap_from_RH(
         thermo_params,
         sim_FT(P_air),
         sim_FT(T_air),
         rh,
+        Thermodynamics.Liquid(),
     )
     return q
 end
@@ -1127,9 +1071,8 @@ horizontal wind speed `u`, specific humidity `q`, and CO2 concentration
 `c_co2`.
 """
 function initialize_drivers(a::PrescribedAtmosphere{FT}, coords) where {FT}
-    keys = (:P_liq, :P_snow, :T, :P, :u, :q, :c_co2, :thermal_state)
-    # The thermal state is a different type
-    types = ([FT for k in keys[1:(end - 1)]]..., Thermodynamics.PhaseEquil{FT})
+    keys = (:P_liq, :P_snow, :T, :P, :u, :q, :c_co2)
+    types = ([FT for k in keys]...,)
     domain_names = ([:surface for k in keys]...,)
     model_name = :drivers
     # intialize_vars packages the variables as a named tuple,
@@ -1185,7 +1128,7 @@ end
     initialize_drivers(r::CoupledAtmosphere{FT}, coords) where {FT}
 
 Creates and returns a NamedTuple for the `CoupledAtmosphere` driver,
-with variables `P_liq`, `P_snow`, `c_co2`, `T`, `P`, `q, `u`, and `thermal_state`.
+with variables `P_liq`, `P_snow`, `c_co2`, `T`, `P`, `q`, and `u`.
 
 This is intended to be used in coupled simulations with ClimaCoupler.jl.
 """
@@ -1193,12 +1136,8 @@ function ClimaLand.initialize_drivers(
     a::CoupledAtmosphere{FT},
     coords,
 ) where {FT}
-    keys = (:P_liq, :P_snow, :c_co2, :T, :P, :q, :u, :thermal_state)
-    types = (
-        [FT for k in 1:(length(keys) - 2)]...,
-        SVector{2, FT},
-        Thermodynamics.PhaseEquil{FT},
-    )
+    keys = (:P_liq, :P_snow, :c_co2, :T, :P, :q, :u)
+    types = ([FT for k in 1:(length(keys) - 1)]..., SVector{2, FT})
     domain_names = ([:surface for k in keys]...,)
     model_name = :drivers
     # intialize_vars packages the variables as a named tuple,
@@ -1282,7 +1221,6 @@ function make_update_drivers(a::PrescribedAtmosphere{FT}) where {FT}
         evaluate!(p.drivers.u, a.u, t)
         evaluate!(p.drivers.q, a.q, t)
         evaluate!(p.drivers.c_co2, a.c_co2, t)
-        set_atmos_ts!(p.drivers.thermal_state, a, p)
     end
     return update_drivers!
 end
@@ -1724,7 +1662,10 @@ function empirical_diffuse_fraction(
     else
         DOY = Dates.dayofyear(start_date + Dates.Second(floor(Int64, t)))
     end
-    RH = ClimaLand.relative_humidity(T, P, q, thermo_params)
+    # Pass explicit zero liquid/ice fractions to avoid Thermodynamics.jl MethodError with default Integer arguments
+    RH = Thermodynamics.relative_humidity(thermo_params, T, P, q) # Note: This used RH over liquid before; not it's RH over ice when below freezing
+    # TODO Replace magic numbers with appropriate constants; for example 1370 probably is TSI and should come from ClimaParams for consistency with rest of model
+    # TODO Nondimensionalize equations appropriately; here we have dimensional constants that are not clearly identified as such, easily leading to errors
     k₀ = FT(1370 * (1 + 0.033 * cos(2π * DOY / 365))) * cosθs
     kₜ = SW_d / k₀
     if kₜ ≤ 0.3