+++ src/standalone/Snow/snow_parameterizations.jl
@@ -132,13 +132,18 @@ Returns the precomputed specific humidity over snow as a weighted
 fraction of the saturated specific humidity over liquid and frozen
 water.
 
-This asumes the atmospheric surface state is stored in p.drivers.
+This uses the atmospheric T, P, q from p.drivers.
 """
 function ClimaLand.component_specific_humidity(model::SnowModel, Y, p)
+    h_sfc = ClimaLand.surface_height(model, Y, p)
+    h_air = model.boundary_conditions.atmos.h
     @. p.snow.q_sfc = snow_surface_specific_humidity(
         p.snow.T_sfc,
         p.snow.q_l,
-        p.drivers.thermal_state,
+        p.drivers.T,
+        p.drivers.P,
+        p.drivers.q,
+        h_air - h_sfc,
         model.parameters,
     )
 
@@ -163,28 +168,38 @@ function ClimaLand.surface_roughness_model(
 end
 
 """
-    snow_surface_specific_humidity(T_sfc::FT, q_l::FT, atmos_ts, parameters) where {FT}
+    snow_surface_specific_humidity(T_sfc::FT, q_l::FT, T_air::FT, P_air::FT, q_air::FT, Δz::FT, parameters) where {FT}
 
 Computes the snow surface specific humidity at a point, assuming a weighted averaged (by mass fraction)
 of the saturated specific humidity over ice and over liquid, at temperature T_sfc.
-
-This approximates the surface air density using an adiabatic approximation and the current atmospheric state.
 """
 function snow_surface_specific_humidity(
     T_sfc::FT,
     q_l::FT,
-    atmos_ts,
+    T_air::FT,
+    P_air::FT,
+    q_air::FT,
+    Δz::FT,
     parameters,
 ) where {FT}
+    surface_flux_params =
+        LP.surface_fluxes_parameters(parameters.earth_param_set)
     thermo_params = LP.thermodynamic_parameters(parameters.earth_param_set)
-    ρ_sfc = ClimaLand.compute_ρ_sfc(thermo_params, atmos_ts, T_sfc)
-    qsat_over_ice = Thermodynamics.q_vap_saturation_generic(
+    ρ_sfc = ClimaLand.compute_ρ_sfc(
+        surface_flux_params,
+        T_air,
+        P_air,
+        q_air,
+        Δz,
+        T_sfc,
+    )
+    qsat_over_ice = Thermodynamics.q_vap_saturation(
         thermo_params,
         T_sfc,
         ρ_sfc,
         Thermodynamics.Ice(),
     )
-    qsat_over_liq = Thermodynamics.q_vap_saturation_generic(
+    qsat_over_liq = Thermodynamics.q_vap_saturation(
         thermo_params,
         T_sfc,
         ρ_sfc,