+++ test/standalone/Snow/snow.jl
@@ -125,19 +125,24 @@ import ClimaLand.Parameters as LP
     )
     thermo_params =
         LP.thermodynamic_parameters(model.parameters.earth_param_set)
-
+    surface_flux_params =
+        LP.surface_fluxes_parameters(model.parameters.earth_param_set)
+    Δh = model.boundary_conditions.atmos.h # h_sfc = 0
     ρ_sfc = @. ClimaLand.compute_ρ_sfc(
-        thermo_params,
-        p.drivers.thermal_state,
+        surface_flux_params,
+        p.drivers.T,
+        p.drivers.P,
+        p.drivers.q,
+        Δh,
         p.snow.T_sfc,
     )
-    q_sfc = @. (1 - p.snow.q_l) * Thermodynamics.q_vap_saturation_generic(
+    q_sfc = @. (1 - p.snow.q_l) * Thermodynamics.q_vap_saturation(
         thermo_params,
         p.snow.T_sfc,
         ρ_sfc,
         Thermodynamics.Ice(),
     ) +
-       p.snow.q_l * Thermodynamics.q_vap_saturation_generic(
+       p.snow.q_l * Thermodynamics.q_vap_saturation(
         thermo_params,
         p.snow.T_sfc,
         ρ_sfc,